moved the current user related funcationality out into a separate folder
