+++ src/core/analytics/SentryTransactionScopeContext.tsx
@@ -1,5 +1,5 @@
 import { createContext, PropsWithChildren, useContext, useEffect, useMemo, useRef } from 'react';
-import { UserMetadata } from '@/domain/community/user/hooks/CurrentUserProvider/useUserMetadataWrapper';
+import { UserMetadata } from '@/domain/community/userCurrent/CurrentUserProvider/useUserMetadataWrapper';
 import { setUserScope, setTransactionScope, TransactionScope } from '@/core/logging/sentry/scope';
 import { useConfig } from '@/domain/platform/config/useConfig';
 import { isEqual, last } from 'lodash';
