+++ src/domain/community/userCurrent/CurrentUserProvider/CurrentUserFull.graphql
